fix registration of loaded plugins
The registration function was turning all plugins into loadable plugins.
This was unnecessary for plugins that are already loaded and directly define
a module field - and could cause downstream bugs.
There happened to not be any visible consequences of this bug,
but I hit some in my AI prompting work, which will come in a separate pull request.
For now, this bug fix produces no changes in behavior.