remove unused actor id to author map
the actor id to author map was originally introduced to support
some features related to showing attributed diffs in editors.
but we don't currently have those features in patchwork.
it was very expensive to maintain the map, it was taking significant
time on every document edit, especially for big docs.
so we are just removing it for now!

note: the timeline sidebar has its own logic for getting author urls
out of changes