Remove service worker init dependency
Problem: The service worker boot sequence currently waits for the URL of
the WASM blob to be sent to it by a connected tab. This means it is
possible for the service worker to be initialized with a wasm blob which
is different to the version of the automerge JS package which wraps it.

Solution: instead of sending the wasm blob URL in an initialization
message, include the base64 encoded wasm blob in the service worker.