if a doc is unavailable, dont crash the whole app (#115)
currently if a doc is unavailable, the whole app crashes.
this commit does the bare minimum fix: if a doc is unavailable,
we make the loading spinner stick around and never finish loading.

this is not the ideal ux; we should ideally distinguish between still loading
and unavailable. however: currently it seems to take 60s to decide that a doc
is unavailable - even if offline! and so in practice, the difference between
"show an error on unavailable" and "show the loading bar permanently" is not
too different.

as a next step to improve this, we should:

- figure out how to detect unavailable faster, rather than waiting 60s
- show unavailable docs differently from loading docs; probably by
  returning a special error value and handling that appropriately.
  (this is a bit annoying with rules of hooks preventing conditional rendering)