avoid loading stale assets in service worker (#119)
We were seeing bugs from loading stale versions of /sdk/dist/*.js.
The cause was this:

When we activate a new service worker, we would fetch fresh
assets to populate a new cache.
But in this process, the _browser cache_ could serve cached older
versions of those assets, which we would then cache in the SW.

This could then cause crashes and other issues stemming from
mismatches between OS code and SDK code.

This commit fixes the issue by forcing network requests from the SW
to avoid the browser cache. This doesn't have significant downside
because we are also caching assets in the service worker anyway,
so it's fine to force loading from network when we're fetching
new versions of assets.

I considered other approaches like adding hashes to sdk built files,
but that seemed more complicated because we'd also need to propagate
those hashes through to the import map.

Testing:

- Confirmed that with a fresh SW we don't see stale assets cached
- The SW still correctly caches assets and serves from its cache
- Offline editing still works